+++ src/helpers/emoji-list.js
