--- src/modules/trip-points/create-element.js
@@ -1,7 +1,0 @@
-// TODO: Не совсем ясна задача данного участка кода
-
-export const createTripPoint = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-  return newElement.firstChild;
-};
